requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
clickclick==20.10.2
connexion==2.14.2
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.0
SQLAlchemy==2.0.20
swagger-ui-bundle==0.0.9
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.2.3
PyJWT~=2.8
pytest~=7.4.1
alembic~=1.12.0
docker~=6.1.3
psycopg~=3.1.12